Remove `c-bar-label` class
The `c-bar-label` class has been a bit of a weird one from the start,
since it can't really live without the `c-bar` class, but also doesn't
live as a child of this one. Since the only important class on this one
is the `mb-3` I feel it makes sense to no longer use it.